# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Santhwanam API X",
    "version": "0.1.0",
    "description": "Minimal OpenAPI spec for IAM (permissions & roles)"
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "description": "API base URL",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:3000",
          "description": "Base URL (e.g., http://localhost:3000 or https://api.example.com)"
        }
      }
    }
  ],
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "description": "JWT access token signed by the server. Token payload contains: sub, userId, authUserId, email, roles"
                    },
                    "refreshToken": {
                      "type": ["string", "null"],
                      "description": "Optional refresh token (from Supabase)"
                    },
                    "expiresAt": {
                      "type": ["integer", "null"],
                      "description": "Optional expiry timestamp (Unix seconds) from Supabase session, if available"
                    },
                    "user": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "userId": { "type": "string" },
                        "email": { "type": "string" },
                        "firstName": { "type": "string" },
                        "lastName": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login. In Postman, set variable: authToken={{token}}"
      }
    }
  }
}